Redis Client Connection

Jedis.Client — Type

Client([; host="127.0.0.1", port=6379, database=0, password="", username=""]) -> Client

Creates a Client instance connecting and authenticating with a Redis host.

Attributes

host::AbstractString: Redis host.

port::Integer: Redis port.

database::Integer: Redis database index.

password::AbstractString: Redis password if any.

username::AbstractString: Redis username if any.

socket::TCPSocket: Socket used for sending and reveiving from Redis host.

is_subscribed::Bool: Whether this Client is actively subscribed to any channels or patterns.

subscriptions::AbstractSet{<:AbstractString}: Set of channels currently subscribed on.

psubscriptions::AbstractSet{<:AbstractString}: Set of patterns currently psubscribed on.

Note

Connection parameters host, port, database, password, username will not change after

client istance is constructed, even with SELECT or CONFIG SET commands.

Examples

julia> client = Client();

julia> set("key", "value"; client=client)
"OK"

julia> get("key"; client=client)
"value"

julia> execute(["DEL", "key"], client)
1

Jedis.execute — Method

execute(command[; client::Client=get_global_client()])

Sends a RESP compliant command to the Redis host and returns the result. The command is either an array of command keywords, or a single command string. Defaults to using the globally set Client.

Examples

julia> execute(["SET", "key", "value"])
"OK"
julia> execute("GET key")
"value"

Jedis.GLOBAL_CLIENT — Constant

GLOBAL_CLIENT = Ref{Client}()

Reference to a global Client object.

Jedis.set_global_client — Function

set_global_client(client::Client)
set_global_client([; host="127.0.0.1", port=6379, database=0, password="", username=""])

Sets a Client object as the GLOBAL_CLIENT[] instance.

Jedis.get_global_client — Function

get_global_client() -> Client

Retrieves the GLOBAL_CLIENT[] instance, if unassigned then initialises it with default values host="127.0.0.1", port=6379, database=0, password="", username="".

Base.copy — Function

copy(client::Client) -> Client

Creates a new Client instance, copying the connection parameters of the input.

Jedis.disconnect! — Function

disconnect!(client::Client)

Closes the client socket connection, it will be rendered unusable.

Jedis.reconnect! — Function

reconnect!(client::Client) -> Client

Reconnects the input client socket connection.

Jedis.wait_until_subscribed — Function

wait_until_subscribed(client::Client)

Blocks until client changes to a subscribed state.

Jedis.wait_until_unsubscribed — Function

wait_until_unsubscribed(client::Client)

Blocks until client changes to a unsubscribed state.

Jedis.wait_until_channel_unsubscribed — Function

wait_until_channel_unsubscribed(client::Client[, channels...])

Blocks until client is unsubscribed from channel(s), leave empty to wait until unsubscribed from all channels.

Jedis.wait_until_pattern_unsubscribed — Function

wait_until_pattern_unsubscribed(client::Client[, patterns...])

Blocks until client is unsubscribed from pattern(s), leave empty to wait until unsubscribed from all patterns.