Redis Commands

Jedis commands all share a common interface, if the client kwarg is not provided then the Jedis.GLOBAL_CLIENT instance will be used:

command(args...; kwargs..., client=get_global_client())

Full list of Jedis commands:

Jedis.auth — Function

auth(password[, username])

Authenticate to the server.

Jedis.select — Function

select(database)

Change the selected database for the current connection.

Jedis.ping — Function

ping()

Ping the server.

Jedis.flushdb — Function

flushdb([; async=false])

Remove all keys from the current database.

Jedis.flushall — Function

flushall([; async=false])

Remove all keys from all databases.

Jedis.quit — Function

quit()

Close the connection.

Jedis.set — Function

set(key, value)

Set the string value of a key.

Base.get — Function

get(key)

Get the value of a key.

Jedis.del — Function

del(key[, keys...])

Delete a key.

Jedis.exists — Function

exists(key[, keys...])

Determine if a key exists.

Jedis.hexists — Function

hexists(key, field)

Determine if a hash field exists.

Base.keys — Function

keys(pattern)

Find all keys matching the pattern.

Jedis.hkeys — Function

hkeys(key)

Get all fields in a hash.

Jedis.setex — Function

setex(key, seconds, value)

Set the value and expiration of a key.

Jedis.expire — Function

expire(key, seconds)

Set a key's tiem to live in seconds.

Jedis.ttl — Function

ttl(key)

Get the time to live for a key.

Jedis.multi — Function

multi()

Mark the start of a transaction block.

Examples

julia> multi()
"OK"

julia> set("key", "value")
"QUEUED"

julia> get("key")
"QUEUED"

julia> exec()
2-element Array{String,1}:
 "OK"
 "value"

Jedis.exec — Function

exec()

Execute all commands issued after MULTI.

Examples

julia> multi()
"OK"

julia> set("key", "value")
"QUEUED"

julia> get("key")
"QUEUED"

julia> exec()
2-element Array{String,1}:
 "OK"
 "value"

Jedis.multi_exec — Function

multi_exec(fn::Function)

Execute a MULTI/EXEC transction in a do block.

Examples

julia> multi_exec() do 
           set("key", "value")
           get("key")
           get("key")
       end
3-element Array{String,1}:
 "OK"
 "value"
 "value"

Jedis.hset — Function

hset(key, field, value)

Set the string value of a hash field.

Jedis.hget — Function

hget(key, field)

Get the value of a hash field.

Jedis.hgetall — Function

hgetall(key)

Get all the fields and values in a hash.

Jedis.hmget — Function

hmget(key, field[, fields...])

Get the values of all the given hash fields.

Jedis.hdel — Function

hdel(key, field[, fields...])

Delete one or more hash fields.

Jedis.lpush — Function

lpush(key, element[, elements...])

Prepend one or multiple elements to a list.

Jedis.rpush — Function

rpush(key, element[, elements...])

Append one or multiple elements to a list.

Jedis.lpos — Function

lpos(key, element[, rank, num_matches, len])

Return the index of matching element on a list.

Jedis.lrem — Function

lrem(key, count, element)

Remove elements from a list.

Jedis.lpop — Function

lpop(key)

Remove and get the first element in a list.

Jedis.rpop — Function

rpop(key)

Remove and get the last element in a list.

Jedis.blpop — Function

blpop(key[, key...; timeout=0])

Remove and get the first element in a list, or block until one is available.

Jedis.brpop — Function

brpop(key[, key...; timeout=0])

Remove and get the last element in a list, or block until one is available.

Jedis.llen — Function

llen(key)

Get the length of a list.

Jedis.lrange — Function

lrange(key, start, stop)

Get a range of elements from a list.

Missing docstring.

Missing docstring for incr. Check Documenter's build log for details.

Missing docstring.

Missing docstring for incrby. Check Documenter's build log for details.

Missing docstring.

Missing docstring for incrbyfloat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hincrby. Check Documenter's build log for details.

Missing docstring.

Missing docstring for hincrbyfloat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for zincrby. Check Documenter's build log for details.